Fix SSH PTY retry policy target ownership
